Big Push
I Brought in ProBuilder and Pro Grids to Make the sheild and Levels soon enough. I Reset the gun to Scale under a parent. I fixed my OCD and Updated the box so that it would line up properly. I Bought a new Material for making grids and brought that in. And the Box prefab Is now organized and Centered. Thats All She wrote :)