Gun and Sheild work
The gun and shield work exactly as expected Pull the left trigger to absorb bullets, or reflecting with the shield calls a script on the bullet that has it speed up and increase damage. The values it increments by can be found on the NormalBullet script. Change them in the bullet prefab